refactor(gsd): extract auto-worktree branch lifecycle
