+++ src/resources/extensions/gsd/auto-worktree.ts
@@ -68,7 +68,11 @@ import { publishMilestone } from "./publication.js";
 import { debugLog } from "./debug-logger.js";
 import { logWarning, logError } from "./workflow-logger.js";
 import {
-  checkoutBranchWithStashGuard,
+  autoWorktreeBranch,
+  fastForwardReusedMilestoneBranchIfSafe,
+  _resolveAutoWorktreeStartPoint,
+} from "./auto-worktree-branch-lifecycle.js";
+import {
   cleanupConflictState,
   gsdJsonlFilesWithConflictMarkers,
   hasConflictMarkers,
@@ -95,12 +99,10 @@ import {
   nativeAddPaths,
   nativeRmForce,
   nativeBranchDelete,
-  nativeBranchForceReset,
   nativeBranchExists,
   nativeDiffNumstat,
   nativeUpdateRef,
   nativeIsAncestor,
-  nativeWorktreeList,
   nativeLsFiles,
 } from "./native-git-bridge.js";
 import {
@@ -113,6 +115,14 @@ import {
 import { gsdHome } from "./gsd-home.js";
 import { type MilestoneScope, type GsdWorkspace, createWorkspace } from "./workspace.js";
 import { WorktreeStateProjection } from "./worktree-state-projection.js";
+
+export {
+  autoWorktreeBranch,
+  enterBranchModeForMilestone,
+  fastForwardReusedMilestoneBranchIfSafe,
+  _isBranchCheckedOutElsewhere,
+  _resolveAutoWorktreeStartPoint,
+} from "./auto-worktree-branch-lifecycle.js";
 
 const PROJECT_PREFERENCES_FILE = "PREFERENCES.md";
 const LEGACY_PROJECT_PREFERENCES_FILE = "preferences.md";
@@ -162,20 +172,6 @@ function isSamePath(a: string, b: string): boolean {
 
 export function _isSamePath(a: string, b: string): boolean {
   return isSamePath(a, b);
-}
-
-export function _resolveAutoWorktreeStartPoint(
-  integrationBranch: string | null | undefined,
-  gitMainBranch: string | null | undefined,
-  branchExists: (branch: string) => boolean,
-): string | undefined {
-  if (integrationBranch) return integrationBranch;
-  return gitMainBranch &&
-    typeof gitMainBranch === "string" &&
-    gitMainBranch.length > 0 &&
-    branchExists(gitMainBranch)
-    ? gitMainBranch
-    : undefined;
 }
 
 function importUntrackedProjectRootContentIntoEmptyWorktree(
@@ -813,171 +809,10 @@ export function syncWorktreeStateBack(
 }
 export { runWorktreePostCreateHook } from "./worktree-post-create-hook.js";
 
-// ─── Auto-Worktree Branch Naming ───────────────────────────────────────────
-
-/** Returns the git branch name for a milestone worktree (`milestone/<MID>`). */
-export function autoWorktreeBranch(milestoneId: string): string {
-  return `milestone/${milestoneId}`;
-}
-
 function normalizeLocalBranchRef(branch: string): string {
   return branch.startsWith("refs/heads/")
     ? branch.slice("refs/heads/".length)
     : branch;
-}
-
-// ─── Branch-mode Entry ─────────────────────────────────────────────────────
-
-/**
- * Enter branch isolation mode for a milestone.
- *
- * Creates `milestone/<MID>` from the integration branch (if it doesn't
- * exist yet) and checks out to it.  No worktree directory is created — the
- * project root is the working copy; only HEAD changes.
- *
- * Uses the same 3-tier integration-branch fallback as createAutoWorktree:
- *   1. META.json recorded integration branch
- *   2. git.main_branch preference
- *   3. nativeDetectMainBranch (origin/HEAD auto-detection)
- */
-export function enterBranchModeForMilestone(
-  basePath: string,
-  milestoneId: string,
-): void {
-  const branch = autoWorktreeBranch(milestoneId);
-  const branchExists = nativeBranchExists(basePath, branch);
-
-  if (!branchExists) {
-    // Create the milestone branch from the integration branch start-point.
-    const integrationBranch =
-      readIntegrationBranch(basePath, milestoneId) ?? undefined;
-    const gitPrefs = loadEffectiveGSDPreferences()?.preferences?.git;
-    const startPoint =
-      _resolveAutoWorktreeStartPoint(
-        integrationBranch,
-        gitPrefs?.main_branch,
-        (branchName) => nativeBranchExists(basePath, branchName),
-      ) ??
-      nativeDetectMainBranch(basePath);
-
-    // TOCTOU ancestry guard (Issue #4980 HIGH-3).
-    //
-    // The outer `branchExists` check at line 1012 is racy: a concurrent
-    // process (parallel-orchestrator worker, side-by-side `gsd` instance,
-    // or manual `git branch` invocation) may have created the branch with
-    // real commits between that check and this point. `nativeBranchForceReset`
-    // does `git branch -f`, which silently overwrites the branch ref —
-    // orphaning any commits not reachable from `startPoint`. Re-check
-    // immediately before the destructive call and refuse if the branch
-    // suddenly exists with non-ancestor commits.
-    //
-    // Note: under single-threaded execution this is rarely reached, but it
-    // is NOT dead code — it is the only barrier against a TOCTOU-induced
-    // commit loss in this code path.
-    const concurrentlyCreated = nativeBranchExists(basePath, branch);
-    if (
-      concurrentlyCreated &&
-      !nativeIsAncestor(basePath, branch, startPoint)
-    ) {
-      throw new GSDError(
-        GSD_GIT_ERROR,
-        `Branch "${branch}" was created concurrently with commits not reachable from "${startPoint}". ` +
-        `Refusing to force-reset — would orphan prior work. ` +
-        `Resume the existing milestone or run \`git branch -D ${branch}\` to discard.`,
-      );
-    }
-    // nativeBranchForceReset creates (or resets) branch at startPoint,
-    // then checkout switches HEAD to it.
-    nativeBranchForceReset(basePath, branch, startPoint);
-    debugLog("auto-worktree", {
-      action: "enterBranchMode",
-      milestoneId,
-      branch,
-      startPoint,
-      created: true,
-    });
-  } else {
-    debugLog("auto-worktree", {
-      action: "enterBranchMode",
-      milestoneId,
-      branch,
-      reused: true,
-    });
-  }
-
-  checkoutBranchWithStashGuard(basePath, branch, `enter-branch-mode:${milestoneId}`);
-}
-
-// ─── Public API ────────────────────────────────────────────────────────────
-
-/**
- * Create a new auto-worktree for a milestone, chdir into it, and store
- * the original base path for later teardown.
- *
- * Atomic: chdir + originalBase update happen in the same try block
- * to prevent split-brain.
- */
-
-/**
- * Forward-merge plan checkbox state from the project root into a freshly
- * re-attached worktree (#778).
- *
- * Phase C: deleted. Writers in workflow-projections.ts, triage-resolution.ts,
- * rule-registry.ts, and auto-post-unit.ts now route through
- * s.canonicalProjectRoot, so non-symlinked worktrees no longer need a local
- * .gsd/ projection — the project-root .gsd/ is the only authoritative source
- * for both reads and writes. copyPlanningArtifacts and reconcilePlanCheckboxes
- * (both formerly here) became dead.
- */
-
-/**
- * True when `branch` is checked out in any worktree listed by
- * `git worktree list --porcelain`. Used to gate ref updates that would
- * otherwise leave a concurrent worktree's HEAD inconsistent with its
- * index/working tree (Codex peer-review of #5538-followup).
- *
- * Best-effort: a `nativeWorktreeList` failure returns true so we err on
- * the side of NOT moving the ref. Better to skip a fast-forward than to
- * silently corrupt another worktree.
- */
-export function _isBranchCheckedOutElsewhere(
-  basePath: string,
-  branch: string,
-): boolean {
-  try {
-    const entries = nativeWorktreeList(basePath);
-    return entries.some((entry) => entry.branch === branch);
-  } catch {
-    return true;
-  }
-}
-
-/**
- * Resolve the integration branch using the same 3-tier fallback as the
- * fresh-create path: META.json → git.main_branch preference → detected
- * main branch. Returns null when no usable target exists.
- */
-function _resolveIntegrationBranchForReuse(
-  basePath: string,
-  milestoneId: string,
-): string | null {
-  const fromMeta = readIntegrationBranch(basePath, milestoneId);
-  if (fromMeta) return fromMeta;
-
-  const gitPrefs = loadEffectiveGSDPreferences()?.preferences?.git;
-  const fromPref = gitPrefs?.main_branch &&
-    typeof gitPrefs.main_branch === "string" &&
-    gitPrefs.main_branch.length > 0 &&
-    nativeBranchExists(basePath, gitPrefs.main_branch)
-    ? gitPrefs.main_branch
-    : null;
-  if (fromPref) return fromPref;
-
-  try {
-    return nativeDetectMainBranch(basePath);
-  } catch {
-    return null;
-  }
 }
 
 function safeCwd(fallback: string): string {
@@ -985,72 +820,6 @@ function safeCwd(fallback: string): string {
     return process.cwd();
   } catch {
     return fallback;
-  }
-}
-
-/**
- * When reusing an existing milestone branch, fast-forward it onto the
- * integration branch when that's safe (branch is a strict ancestor of
- * integration — no commits would be lost). Skips when the branch has its
- * own commits ahead of integration, when the integration branch can't be
- * resolved, or when any git operation fails — the merge gate at milestone
- * completion will surface real divergence as a conflict.
- *
- * The previous behavior re-attached the worktree to whatever stale tip
- * the branch held, which caused new milestone work to fork from a base
- * missing prior milestones' merges (#5538-followup).
- */
-export function fastForwardReusedMilestoneBranchIfSafe(
-  basePath: string,
-  milestoneId: string,
-  branch: string,
-): void {
-  try {
-    const integrationBranch = _resolveIntegrationBranchForReuse(basePath, milestoneId);
-    if (!integrationBranch || integrationBranch === branch) return;
-    if (!nativeBranchExists(basePath, integrationBranch)) return;
-
-    // Pure fast-forward only: branch must be a strict ancestor of integration.
-    // If the branch has its own commits ahead, leave it alone.
-    if (!nativeIsAncestor(basePath, branch, integrationBranch)) {
-      debugLog("createAutoWorktree", {
-        phase: "skip-ff-branch-not-ancestor",
-        milestoneId,
-        branch,
-        integration: integrationBranch,
-      });
-      return;
-    }
-
-    // Codex peer-review: `nativeUpdateRef` succeeds even when the branch is
-    // currently checked out in another worktree, leaving that worktree's HEAD
-    // inconsistent with its index/work tree. Skip the fast-forward if any
-    // listed worktree has this branch checked out — the merge gate at
-    // milestone-completion will surface stale-base divergence as a conflict
-    // instead of silently corrupting the other worktree's state.
-    if (_isBranchCheckedOutElsewhere(basePath, branch)) {
-      debugLog("createAutoWorktree", {
-        phase: "skip-ff-branch-checked-out-elsewhere",
-        milestoneId,
-        branch,
-      });
-      return;
-    }
-
-    nativeUpdateRef(basePath, `refs/heads/${branch}`, integrationBranch);
-    debugLog("createAutoWorktree", {
-      phase: "fast-forward-reused-branch",
-      milestoneId,
-      branch,
-      integration: integrationBranch,
-    });
-  } catch (err) {
-    debugLog("createAutoWorktree", {
-      phase: "fast-forward-reused-branch-failed",
-      milestoneId,
-      branch,
-      error: err instanceof Error ? err.message : String(err),
-    });
   }
 }
